Don't update modifications if no initial value
This happens with hidden properties of plugins which are not used and
therefore there's no value on clawsrc. Could also happen with removed
hidden properties which are not present in newer Claws Mail installs.

In both cases the widgets are disabled in UI, but the handler is run on
startup, causing a warning (undefined value used in comparison) and
modifiying the counter, wich in turn causes the warning dialog to be
wrongly shown on exit.