Fix RFE 4051: enable sortable columns on hotkeys
Unfortunately the hidden column which defines oddity is also sorted
along the visible column, so the alternating background color on rows is
lost. The observed effect now is to be alternating more randomly. Not
too bad, but not the ideal.

Fixing this requires more investigation, because the oddity values can't
be swapped within the sort function iself, it segfaults after some
messages: Deep recursion on subroutine "Clawsker::store_sort"…